requirements.txt
certifi==2020.12.5
click==8.0.3
Flask==2.0.2
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
